Implemented new version of k-means for inner product spaces -- presumably only good for inputs with close to uniform norms
It does not use a transformation to L2 or cosine. To compute the new centroid of a cluster,
we sum its normalized vectors (as directions) and then rescale the centroid vector to the average
so that its norm is the average of its vector's norms. We should also test whether
normalizing the directions is helpful or whether it should be discarded.